==> notebook/1_llamada_API.ipynb <==
# Importamos pandas
import pandas as pd

# Importamos las funciones de soporte
import sys
sys.path.append("../")
import src.soporte as sp

# Realizo llamadas a la API REData para obtener el balance
sp.balance()

# Cargamos los resultados para obtener una pequeña visualización de los datos
df_balance = pd.read_csv("../data/scrap/balance.csv")
df_balance.head()

# Realizo llamadas a la API REData para obtener el balance por Comunidad Autónoma
sp.balance_ccaa()

# Realizo llamadas a la API REData para obtener la demanda de energía
sp.demanda()

# Realizo llamadas a la API REData para obtener la demanda de energía por Comunidad Autónoma
sp.demanda_ccaa()

# Realizo llamadas a la API REData para obtener la demanda diaria máxima de cada mes
sp.demanda_max_diaria()

# Realizo llamadas a la API REData para obtener la demanda horaria máxima de cada mes
sp.demanda_max_horaria()

# Realizo llamadas a la API REData para obtener las pérdidas energéticas por transporte
sp.perdidas_transporte()

# Realizo llamadas a la API REData para obtener la potencia máxima mensual
sp.potencia_maxima_instantanea()

# Realizo llamadas a la API REData para obtener la demanda de energía real por minutos, además de las predicciones españolas y la energía programada
sp.demanda_tiempo_real()

# Realizo llamadas a la API REData para obtener los datos sobre la evolución de las energías renovables
sp.evolucion_renovable()

# Realizo llamadas a la API REData para obtener los datos sobre la evolución de las energías renovables por Comunidades Autónomas
sp.evolucion_renovable_ccaa()

# Realizo llamadas a la API REData para obtener Los porcentajes procedentes de energías renovables
sp.estructura_renovable()

# Realizo llamadas a la API REData para obtener los datos sobre la cantidades de emisión de CO2
sp.emisiones_CO2()

# Realizo llamadas a la API REData para obtener los datos sobre la evolución del precio de la energía
sp.precios_mercados()
