==> notebook/3_Serie_temporal.ipynb <==
import pandas as pd
from pycaret.time_series import *
import datetime

import matplotlib.pyplot as plt
plt.rcParams['figure.figsize'] = [10, 5]

import warnings
warnings.filterwarnings("ignore")

# Cargar datos desde un archivo CSV
df = pd.read_csv("../data/scrap/demanda_tiempo_real.csv")

# Eliminar columnas innecesarias del DataFrame
df.drop(["percentage_Demanda real", "percentage_Demanda programada", "percentage_Demanda prevista"], axis=1, inplace=True)

# Convertir la columna "datetime" a tipo de dato datetime con zona horaria UTC y formato específico
df["datetime"] = pd.to_datetime(df["datetime"], utc=True, format='%Y-%m-%d %H:%M:%S%z')

# Crear una nueva columna "drop" que contiene "datetime" si los minutos no son múltiplo de 10, de lo contrario contiene "drop"
df["drop"] = df["datetime"].where(df["datetime"].dt.minute % 10 == 0, other="drop")

# Filtrar las filas donde "drop" no sea igual a "drop", eliminando las filas que no son múltiplo de 10 minutos
df = df[df["drop"] != "drop"]

# Eliminar la columna "drop" ya que ya no es necesaria
df.drop(["drop"], inplace=True, axis=1)

# Seleccionar y renombrar las columnas relevantes
df = df[["datetime", "value_Demanda real"]].rename(columns={"value_Demanda real": "demanda_real"})


# Crear una nueva columna "datetime" con una secuencia de fechas y horas
df["datetime"] = pd.date_range(datetime(2013, 12, 31, hour = 23, minute=0), periods = 486228, freq = '10min')

df.head()

# Convertir la columna "datetime" a tipo de dato datetime y asignarla como índice del DataFrame
df.index = pd.to_datetime(df["datetime"])

# Inferir la frecuencia de los índices del DataFrame
df.index.freq = pd.infer_freq(df.index)
df.index

# Eliminar la columna 'datetime' del DataFrame 'df' de forma inplace
df.drop(columns=['datetime'], inplace=True)
df.head()

# Visualizamos cuantos nulos hay
df.isnull().sum()

# Renombramos la columna a "valores"
df.columns = ["valores"]

# Preparamos los settings de Pycaret
s = setup(df, fh = 3, fold = 5, session_id = 42)

# Buscamos y comparamos los mejores modelos
best = compare_models()

# Visualizamos nuestro mejor modelo
best

# Entrenamos un modelo con las caraterísticas del mejor modelo
final_best = finalize_model(best)


# Guardamos nuestro modelo que tanto nos ha costado entrenar
save_model(final_best, '../data/modelo/my_final_best_model')

# Realizamos una predicción de 24 valores a partir de los datos con los que ha sido entrenados
predictions = predict_model(best, fh = 24)

# Guardamos la predicción para futura visualización
predictions.to_csv("../data/visualizacion/predicción.csv")

# Podemos cargar el modelo que hemos entrenado en esta ruta
loaded_model = load_model('../data/modelo/my_final_best_model')
