==> notebook/4_Prediccion.ipynb <==
import pandas as pd
import pickle

# Cargamos nuestros datos
predicción = pd.read_csv("../data/visualizacion/predicción.csv", index_col = 0)
with open(f"../data/visualizacion/px_demanda_real.pkl", "rb") as demanda_real:
    demanda = pickle.load(demanda_real)

# Resetear el índice del DataFrame de predicciones y renombrar la columna del índice como "Fecha"
predicción_demanda = predicción.reset_index().rename(columns = {"index": "Fecha"})

# Convertir la columna "Fecha" a tipo de dato datetime con zona horaria en UTC
predicción_demanda["Fecha"] = pd.to_datetime(predicción_demanda["Fecha"], utc = True, format = "%Y-%m-%d %H:%M:%S%z")

# Convertir la zona horaria de la columna "Fecha" a "Europe/Paris"
predicción_demanda["Fecha"] = predicción_demanda["Fecha"].dt.tz_convert("Europe/Paris")

# Renombrar la columna "y_pred" como "Prediccion propia en MW"
predicción_demanda.rename(columns = {"y_pred": "Prediccion propia en MW"}, inplace = True)

# Mostrar el DataFrame de predicciones
predicción_demanda


# Crear una nueva columna "drop" en el DataFrame "demanda" con los valores de la columna "Fecha" cuando los minutos no son múltiplos de 10, o "drop" en caso contrario
demanda["drop"] = demanda["Fecha"].where(demanda["Fecha"].dt.minute % 10 == 0, other = "drop")

# Filtrar el DataFrame "demanda" eliminando las filas donde la columna "drop" tiene el valor "drop"
demanda = demanda[demanda["drop"] != "drop"]

# Seleccionar solo las columnas "Fecha", "Demanda real en MW" y "Demanda prevista en MW" del DataFrame "demanda"
demanda = demanda[["Fecha", "Demanda real en MW", "Demanda prevista en MW"]]

# Definir las fechas de inicio y final para el rango de datos deseado
inicio = "2023-03-31 14:30:00+02:00"
final = "2023-03-31 18:20:00+02:00"

# Filtrar el DataFrame "demanda" para obtener solo las filas cuyos valores en la columna "Fecha" estén dentro del rango definido por las fechas de inicio y final
df_demanda = demanda[(demanda["Fecha"] >= inicio) & (demanda["Fecha"] <= final)]

df_demanda.head()

# Combinar (merge) los DataFrames "predicción_demanda" y "df_demanda" basándose en una columna en común, que en este caso es la columna "Fecha"
df_demanda_predicción = pd.merge(predicción_demanda, df_demanda)

df_demanda_predicción

# Guardamos los datos para posterior visualización
with open("../data/visualizacion/px_prediccion.pkl", "wb") as prediccion_demanda:
    pickle.dump(df_demanda_predicción, prediccion_demanda)
